pyproject.toml
[tool.isort]

profile = "black"

src_paths = ["isort", "test"]


[tool.poetry]
name = "Html Generator"
version = "0.0.1"
description = "it will generate webpage for goods sells"
authors = ["natnael"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
Jinja2 = "^3.0.3"
Flask = "^2.0.3"
google-cloud-storage = "^2.2.1"
google-cloud-pubsub = "^2.12.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
wemake-python-styleguide = "^0.16.0"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
